renamed package to woningwaardering
