+++ pyproject.toml
@@ -3,9 +3,9 @@ requires = ["setuptools>=61.0.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "woningwaarderingsstelsel"
+name = "woningwaardering"
 version = "0.1.0"
-description = "Berekent de punten van een woning m.b.v. het woningwaarderingstelsel."
+description = "Berekent de punten van een woning op basis van het woningwaarderingstelsel."
 authors = [
     { name = "Ben Verhees", email = "ben.verhees@woonstadrotterdam.nl" },
     { name = "Tiddo Loos", email = "tiddo.loos@woonstadrotterdam.nl" },
